requirements.txt
django
djangorestframework==3.13.1
django-cors-headers==3.11.0

# scipy
# numpy
# matplotlib
# pandas
# sklearn
# alpha_vantage
# keras
# tensorflow
# mplfinance

# jupyterlab
# jupyter
# notebook
